NPAL16033: Pb with print when the list is too long
setText method modified. Previously exceptions occurred if one of paragraphs of the input string "s" is too long.  Now long paragraph is limited with threshold numbers of characters and finished by " ..." string. Note that length of the first paragraph of the string is verified only because it is enough for bug fixing. If it will be insufficient for other cases then more complicated check should be implemented. At present it is not done because of possible performance problem.

Bug NPAL16035: Pb with the history list of salome (GUI)

Now in python console after use of history of commands view is scrolled at the left side of console if current command can be completely displayed on the screen.

void scrollViewAfterHistoryUsing( const QString& command ); method added